# Improve the speed of tracking arrays declared over a domain