--- a/test/modules/sungeun/init/printModuleInitOrder.good
+++ b/test/modules/sungeun/init/printModuleInitOrder.good
@@ -33,7 +33,6 @@ Initializing Modules:
          ChapelHashtable
         ExternalArray
         RangeChunk
-        ChapelHashing
        ChapelNumLocales
        Sys
       LocaleModelHelpRuntime