--- a/test/modules/sungeun/init/printModuleInitOrder.na-none.good
+++ b/test/modules/sungeun/init/printModuleInitOrder.na-none.good
@@ -32,7 +32,6 @@ Initializing Modules:
          ChapelHashtable
         ExternalArray
         RangeChunk
-        ChapelHashing
        ChapelNumLocales
        Sys
       LocaleModelHelpRuntime