--- a/test/optimizations/constDomain/trackArraysFlag.chpl
+++ b/test/optimizations/constDomain/trackArraysFlag.chpl
@@ -2,4 +2,4 @@ const d = {1..3};
 
 var a: [d] int;
 
-writeln("The domain is tracking ", d._value._arrs.size, " arrays");
+writeln("The domain is tracking ", +reduce for d._value._arrs do 1, " arrays");