--- a/test/optimizations/deadCodeElimination/elliot/countDeadModules.comm-ofi.good
+++ b/test/optimizations/deadCodeElimination/elliot/countDeadModules.comm-ofi.good
@@ -1,1 +1,1 @@
-Removed 24 dead modules.
+Removed 25 dead modules.