--- a/test/optimizations/deadCodeElimination/elliot/countDeadModules.comm-ugni.good
+++ b/test/optimizations/deadCodeElimination/elliot/countDeadModules.comm-ugni.good
@@ -1,1 +1,1 @@
-Removed 24 dead modules.
+Removed 25 dead modules.